refactor: component 이름 변경, import type 경로 변경
