+++ frontend/src/components/Submission/Connecting.tsx
@@ -1,0 +1,24 @@
+import { css } from '@style/css';
+
+import Loading from '@/components/Common/Loading';
+
+interface Props {
+  isConnected: boolean;
+}
+
+export default function Connecting(props: Props) {
+  return (
+    <>
+      {props.isConnected && (
+        <div className={rowStyle}>
+          <span>연결 중...</span>
+          <Loading color="darkorange" size="24px" />
+        </div>
+      )}
+    </>
+  );
+}
+const rowStyle = css({
+  display: 'flex',
+  gap: '0.5rem',
+});
